requirements.txt
Pillow==9.3.0
pyzbar==0.1.10
qrcode==7.4.2
